Extract search item to its own component file
search_live was > 900 lines long, and we might look to use this in other
places, e.g. if we pursue #1192
